Added final version

+++ ds_management_app/bin/ds_reload.py
@@ -1,19 +1,20 @@
 # ds_reload.py
 import splunk.Intersplunk
-import csv, os,platform
-import shutil, time
+import csv, os,platform, traceback
+import shutil, time, sys
 from concurrent.futures import ThreadPoolExecutor
 from multiprocessing import cpu_count
+from splunklib.searchcommands import dispatch, GeneratingCommand, Configuration, Option
 from ds_utils import log,get_apps_for_input,get_apps_checkpoint
-from setup import compress_app_update_checkpoint,csv_file, create_machine_types_filter_file
+from setup import compress_app_update_checkpoint,serverclass_csv_file, create_machine_types_filter_file
 from ds_utils import dc_info_csv
 from extract_csv_parms import extrace_csv
 from splunk.clilib.bundle_paths import make_splunkhome_path
 import splunk.appserver.mrsparkle.lib.util as splunk_lib_util
 
-apps_download_list_dir = splunk_lib_util.make_splunkhome_path(['etc', 'apps', 'ds_management_app', 'appserver', 'static','apps_download_list'])
-temp_apps_download_list_dir = splunk_lib_util.make_splunkhome_path(['etc', 'apps', 'ds_management_app', 'var', 'run','temp_apps_download_list'])
-temp_dc_info_csv = splunk_lib_util.make_splunkhome_path(['etc', 'apps', 'ds_management_app', 'var', 'run', 'dc_info_csv.csv'])
+apps_download_list_dir = splunk_lib_util.make_splunkhome_path(['etc', 'system', 'static', 'ds_management_app','apps_download_list'])
+temp_apps_download_list_dir = splunk_lib_util.make_splunkhome_path(['var','run','ds_management_app','temp_apps_download_list'])
+temp_dc_info_csv = splunk_lib_util.make_splunkhome_path(['var','run','ds_management_app', 'lookups','dc_info_csv.csv'])
 
 RELOAD_SUMMARY=[]
 
@@ -41,7 +42,7 @@ def process_row(row):
     try:
         if is_within_last_24_hours(int(row["_time"])):
             required_unique_keys = [row["guid"], row["ip"], row["hostname"], row["servername"]]
-            required_apps = get_apps_for_input(required_unique_keys, csv_file, row['os'])
+            required_apps = get_apps_for_input(required_unique_keys, serverclass_csv_file, row['os'],row["guid"])
             apps_with_checkpoint = get_apps_checkpoint(list(required_apps))
 
             make_name = f"{row['guid']}__{row['private_ip']}__{row['hostname']}__{row['os']}.txt"
@@ -62,7 +63,6 @@ def process_row(row):
 def copy_files_to_tmp_location():
     
     try:
-    
         # Copy the CSV file to the temporary location
         shutil.copy(dc_info_csv, temp_dc_info_csv)
         # log("INFO", f"Copied {dc_info_csv} to {temp_dc_info_csv}")
@@ -87,7 +87,10 @@ def copy_files_to_tmp_location():
         os.makedirs(apps_download_list_dir, exist_ok=True)
 
         for file_name in os.listdir(temp_apps_download_list_dir):
-            shutil.move(os.path.join(temp_apps_download_list_dir, file_name), apps_download_list_dir)
+            if os.path.exists(os.path.join(apps_download_list_dir, file_name)):
+                continue
+            else:
+                shutil.move(os.path.join(temp_apps_download_list_dir, file_name), apps_download_list_dir)
         
     except Exception as e:
         log("ERROR", "Error in deployment server reload")
@@ -101,20 +104,28 @@ def copy_files_to_tmp_location():
         if os.path.exists(temp_apps_download_list_dir):
             shutil.rmtree(temp_apps_download_list_dir)
 
-def main():
-    try:
-        log("INFO", "Reload deployment server is in progress...")
-        compress_app_update_checkpoint()
-        copy_files_to_tmp_location()
-        extrace_csv()
-        create_machine_types_filter_file()
-        splunk.Intersplunk.outputResults([{"status":"success", "_raw": "Successfully reloaded deployment server","summary":RELOAD_SUMMARY }])
-    
-    except Exception as e:
-        log("ERROR", "Error in deployment server reload")
-        log("ERROR", str(e))
-        splunk.Intersplunk.generateErrorResults([{"status":"error", "_raw": f'Error: {str(e)}', "summary":RELOAD_SUMMARY }])
+@Configuration()
+class ReloadDS(GeneratingCommand):
+    dsIP = Option(require=False)
+    repositoryLocation=Option(require=False)
+    phonehome=Option(require=False)
+    pullDelay=Option(require=False)
 
-            
-if __name__ == '__main__':
-    main()
+    def generate(self):
+        try:
+            log("INFO", "Reload deployment server is in progress...")
+            compress_app_update_checkpoint()
+            extrace_csv()
+            create_machine_types_filter_file()
+            copy_files_to_tmp_location()
+            log("INFO", "Successfully reloaded deployment server !!!")
+            yield {"status":"success", "_raw": "Successfully reloaded deployment server","summary":RELOAD_SUMMARY }
+        
+        except Exception as e:
+            log("ERROR", "Error in deployment server reload")
+            log("ERROR",traceback.format_exc())
+            log("ERROR", str(e))
+            yield {"status":"error", "_raw": f'Error: {str(e)}', "summary":RELOAD_SUMMARY }
+
+
+dispatch(ReloadDS, sys.argv, sys.stdin, sys.stdout, __name__)
